Bugfixes Ich hate this messages ...

+++ actions/delete-action.php
@@ -1,19 +1,28 @@
 <?php
 
-//AUTH-Detection!!!
-require_once "../require/mysql.php";
+if (isset($_SESSION['authACP'])) {
+    if ($_SESSION['authACP'] == 1) {
+
+        require_once "../require/mysql.php";
 
 
-if (isset($_GET["job"]) && isset($_GET["name"])) {#
-    $job = $_GET["job"];
-    $name = $_GET["name"];
+        if (isset($_GET["job"]) && isset($_GET["name"])) {#
+            $job = $_GET["job"];
+            $name = $_GET["name"];
 
-    $sql = "DELETE FROM members WHERE name = '$name' and job = '$job' LIMIT 1";
+            $sql = "DELETE FROM members WHERE name = '$name' and job = '$job' LIMIT 1";
 
-    echo $sql;
+            echo $sql;
 
-    $mysqli->query($sql) or die($mysqli->error);
-    header("Location: ../acp.php");
+            $mysqli->query($sql) or die($mysqli->error);
+            header("Location: ../acp.php");
+        } else {
+            header("Location: ../index.php");
+        }
+
+    } else {
+        header("Location: ../index.php");
+    }
 } else {
     header("Location: ../index.php");
 }
